Mac mini 2016-08-15

+++ Everywhere/EverywhereVisualViewControllers/AssetDetailVC.m
@@ -54,7 +54,7 @@
             playButton.hidden = YES;
         }
         
-        if (currentIndex > 0) noteLabel.text = [NSString stringWithFormat:@"%ld/%lu",currentIndex + 1,(unsigned long)assetArray.count];
+        if (currentIndex > 0) noteLabel.text = [NSString stringWithFormat:@"%lu/%lu",(unsigned long)(currentIndex + 1),(unsigned long)assetArray.count];
         else noteLabel.text = NSLocalizedString(@"Swipe up to quite", @"上滑退出");
     }
 }
